add previous-nuget source to the differ
